+++ .solcover.js
@@ -1,0 +1,7 @@
+module.exports = {
+  port: 8545,
+  norpc: false,
+  compileCommand: 'truffle compile --all',
+  testCommand: 'truffle test --network coverage',
+  skipFiles: ['AbstractChain.sol']
+}