+++ truffle.js
@@ -18,6 +18,13 @@ module.exports = {
       host: '10.0.0.15',
       port: 8545,
       network_id: 1
-    }
+    },
+    coverage: {
+      host: "localhost",
+      network_id: "*",
+      port: 8545,         // <-- If you change this, also set the port option in .solcover.js.
+      gas: 0xfffffffffff, // <-- Use this high gas value
+      gasPrice: 0x01      // <-- Use this low gas price
+    },
   }
 };